Move CustomReqparser to /utils
