+++ python-backend/app/api/parties/party_appt/resources/mine_party_appt_admin_resource.py
@@ -3,9 +3,9 @@ from flask_restplus import Resource
 
 from app.extensions import api
 from ..models.mine_party_appt import MinePartyAppointment
-from ...custom_reqparser import CustomReqparser
 from ....utils.access_decorators import requires_role_mine_admin
 from ....utils.resources_mixins import UserMixin, ErrorMixin
+from app.api.utils.custom_reqparser import CustomReqparser
 
 
 class MinePartyApptAdminResource(Resource, UserMixin, ErrorMixin):
