+++ python-backend/tests/mines/variances/resources/test_variance_list_resource.py
@@ -2,7 +2,7 @@ import json
 
 from tests.factories import VarianceFactory, MineFactory
 from tests.status_code_gen import RandomComplianceArticleId
-from app.api.parties.custom_reqparser import DEFAULT_MISSING_REQUIRED
+from app.api.utils.custom_reqparser import DEFAULT_MISSING_REQUIRED
 
 
 # GET
